refactor : PageResponse의 totalElement 타입 변경

+++ src/main/java/team7/hrbank/domain/employee/repository/CustomEmployeeRepository.java
@@ -13,7 +13,7 @@ public interface CustomEmployeeRepository {
   List<Employee> findEmployees(EmployeeFindRequest employeeFindRequest);
 
   // 총 직원 수 집계
-  Integer totalCountEmployee(EmployeeCountRequest request);
+  long totalCountEmployee(EmployeeCountRequest request);
 
   // 해당 년도에 입사한 직원 중 가장 마지막에 만들어진 직원의 사원번호
   String selectLatestEmployeeNumberByHireDateYear(int year);
